init RemoteRender
